refactor ray runner
