+++ python/orca/src/bigdl/orca/learn/pytorch/core/__init__.py
@@ -17,3 +17,4 @@
 
 # from .base_ray_estimator import BaseRayEstimator
 from .base_runner import BaseRunner
+from .base_ray_estimator import BaseRayEstimator
